image component definitions
