+++ src/routes/patterns/index.js
@@ -12,6 +12,7 @@ import {
 import Buttons from "./components/Buttons";
 import Layout from "./components/Layout";
 import Icons from "./components/Icons";
+import Images from "./components/Images";
 import Loaders from "./components/Loaders";
 import Overlays from "./components/Overlays";
 import Typography from "./components/Typography";
@@ -22,9 +23,6 @@ export default class PatternLibrary extends Component {
       <Section>
         <NavBar scroll>
           <NavItem fit>
-            <Button link>Layout</Button>
-          </NavItem>
-          <NavItem fit>
             <Button link>Typography</Button>
           </NavItem>
           <NavItem fit>
@@ -34,10 +32,16 @@ export default class PatternLibrary extends Component {
             <Button link>Forms</Button>
           </NavItem>
           <NavItem fit>
-            <Button link>Overlays</Button>
+            <Button link>Layout</Button>
+          </NavItem>
+          <NavItem fit>
+            <Button link>Images</Button>
           </NavItem>
           <NavItem fit>
             <Button link>Icons</Button>
+          </NavItem>
+          <NavItem fit>
+            <Button link>Overlays</Button>
           </NavItem>
           <NavItem fit>
             <Button link>Loader</Button>
@@ -49,13 +53,13 @@ export default class PatternLibrary extends Component {
           <Paragraph large mb={48}>
             These components are a collection of elements that can be reused across the Anni interface.
           </Paragraph>
-          <Layout />
           <Typography />
           <Buttons />
-          <Overlays />
+          <Layout />
+          <Images />
           <Icons />
+          <Overlays />
           <Loaders />
-          {/* <FormFields /> */}
           {/* <Panels /> */}
         </Content>
       </Section>
